+++ src/main/java/ongi/ongibe/domain/album/UserAlbumRole.java
@@ -1,4 +1,4 @@
-package ongi.ongibe;
+package ongi.ongibe.domain.album;
 
 public enum UserAlbumRole {
     OWNER, NORMAL
